requirements.txt
Django==1.9
django-treebeard==3.0
django-widget-tweaks==1.4.1
Jinja2==2.8
Pillow==3.0.0
six==1.10.0
pytest==2.8.5
pytest-cov==2.2.0
pytest-django==2.9.1

